Create UWPFolder class

+++ HamburgerUI/Services/FileListFactory.cs
@@ -25,12 +25,12 @@ namespace HamburgerUI.Services
             {
                 var folder = new DirectoryInfo(@path);
                 var fileList = new List<IFile>();
-
+                                
                 try
                 {
                     foreach (var fi in folder.EnumerateFiles())
                     {
-                        var tempFile = new Models.File(fi.Name, fi.CreationTime, fi.FullName.Substring(2), fi.Length, fi.Extension);
+                        var tempFile = new Models.File(fi.Name, fi.CreationTime, fi.FullName.Substring(2), (ulong)fi.Length, fi.Extension);
                         fileList.Add(tempFile);
                     }
                 }
